requirements.txt
python-binance
python-dotenv
rich
questionary
streamlit
portalocker
flask
fastapi 
uvicorn
pytest 
pytest-cov
pytest-asyncio
pyarrow
requests
pandas_ta
anthropic
